An implementation of the BlitzMail protocol
Starting in around 1987, Dartmouth College in Hanover, NH, USA developed its
own home-grown e-mail system called BlitzMail.  Although it was later bypassed
by standards and convention, and eventually replaced, it had some nice features
that weren't available in other e-mail programs.

BlitzMail is a client-server architecture, in which each user's e-mail is
stored on a central server and accessed through client software that speaks the
BlitzMail protocol (as well as POP and IMAP).  This library is a Python
implementation of the native protocol.